Calculate Matrix Column Variance

Computes the sample variance for each column of a numeric matrix

Usage

col_vars(mat, cores = 1)

Arguments

mat

A numeric matrix.

cores

Number of cores to use for parallel computation. Defaults to 1.

Value

A named numeric vector of column variances

Details

Variances for columns with one unique value after dropping NA are set to NA.

Examples

col_vars(t(khanmiss1))
#>         g1         g2         g3         g4         g5         g6         g7 
#> 414020.949 148453.125 454283.693 154292.401 110384.778 487724.184   1456.120 
#>         g8         g9        g10        g11        g12        g13        g14 
#> 545124.433 596495.369 396556.812   1049.057 417036.074  59763.607 372644.765 
#>        g15        g16        g17        g18        g19        g20        g21 
#> 508301.539 307078.821 476103.243 133826.113 408165.709 477510.636 478854.092 
#>        g22        g23        g24        g25        g26        g27        g28 
#>   1008.397 176191.415 107270.562 284525.479   1430.286 451184.350 623728.466 
#>        g29        g30        g31        g32        g33        g34        g35 
#> 178425.338 575578.055 168633.571 534687.894 220937.620 198592.737  70320.813 
#>        g36        g37        g38        g39        g40        g41        g42 
#> 415029.792 321272.231 203725.199 155783.967 117931.909 276897.361  14079.348 
#>        g43        g44        g45        g46        g47        g48        g49 
#> 209498.338 192128.904 428429.381 467700.372  14745.716 489045.891 166438.913 
#>        g50        g51        g52        g53        g54        g55        g56 
#> 124656.171  61632.257 173900.285 255576.458 127251.530 271009.055 538629.683 
#>        g57        g58        g59        g60        g61        g62        g63 
#>   6996.845 120311.166  70011.842   5718.745   4021.150  86383.048 382060.415 
#>        g64        g65        g66        g67        g68        g69        g70 
#> 541432.209 505572.747 217644.060 164150.678 433244.403 156522.369  89849.534 
#>        g71        g72        g73        g74        g75        g76        g77 
#> 227942.074 286687.842  86749.655 217628.179 181627.401 510420.286 422550.236 
#>        g78        g79        g80        g81        g82        g83        g84 
#> 113016.440  46998.587 161282.096 129846.564  29281.369  72362.405 376362.320 
#>        g85        g86        g87        g88        g89        g90        g91 
#> 274704.871 110142.401 310164.128 328065.419 142576.792 189399.510 128685.143 
#>        g92        g93        g94        g95        g96        g97        g98 
#> 455731.704  88412.844  49849.930 136124.733 127709.113 191069.337  65395.295 
#>        g99       g100       g101       g102       g103       g104       g105 
#> 525496.607 108357.985 552928.661 225670.822 116560.961 406873.865  94514.429 
#>       g106       g107       g108       g109       g110       g111       g112 
#> 121587.566 367573.754 126136.961  94915.465  23746.040  65578.156 512230.231 
#>       g113       g114       g115       g116       g117       g118       g119 
#>  62179.601 449096.672 155836.595  90963.811  41522.401 537867.392 244807.762 
#>       g120       g121       g122       g123       g124       g125       g126 
#> 565152.350 180587.606 292130.184 152109.188 540236.784 208025.111 422941.491 
#>       g127       g128       g129       g130       g131       g132       g133 
#> 364325.382 446290.414 191148.571 535608.669 240884.683 118703.124 206917.988 
#>       g134       g135       g136       g137       g138       g139       g140 
#> 163455.097 441318.193 480720.666 443142.758 303895.544 178858.375 481697.323 
#>       g141       g142       g143       g144       g145       g146       g147 
#> 585146.967 335697.304 153955.447 205341.446  93680.281 399498.501  29669.079 
#>       g148       g149       g150       g151       g152       g153       g154 
#>   1573.956 428316.930 137825.231 459739.604 581723.242 426530.254 120773.348 
#>       g155       g156       g157       g158       g159       g160       g161 
#> 124881.894 338275.220 161983.175 330937.289 400492.908  65960.403 237414.401 
#>       g162       g163       g164       g165       g166       g167       g168 
#> 424218.876  65689.498 403200.355 287122.322 270988.225 129209.107  69780.060 
#>       g169       g170       g171       g172       g173       g174       g175 
#> 319270.953 293739.949 377252.329  91811.078  82596.813 155831.128 212464.984 
#>       g176       g177       g178       g179       g180       g181       g182 
#> 100187.475  33106.285 285777.676 397577.645  17443.097 304273.991 196638.730 
#>       g183       g184       g185       g186       g187       g188       g189 
#> 511822.252  26673.887 353910.917 147810.378 289074.754 228060.060 488225.054 
#>       g190       g191       g192       g193       g194       g195       g196 
#> 116871.059 111186.673 282621.627 178994.005 246451.512 212692.146 386274.157 
#>       g197       g198       g199       g200       g201       g202       g203 
#>  46845.088 215463.537 126233.526 406736.037 197035.693  15058.520 247933.961 
#>       g204       g205       g206       g207       g208       g209       g210 
#> 165659.856 494467.569 547947.275 545399.059 138636.705 189576.870 143906.074 
#>       g211       g212       g213       g214       g215       g216       g217 
#> 400557.931 465420.209 177320.207  71848.355  39539.382 227981.313 173831.586 
#>       g218       g219       g220       g221       g222       g223       g224 
#>  57146.618 145606.189  79612.907 325301.048 362488.666 308422.369  70837.703 
#>       g225       g226       g227       g228       g229       g230       g231 
#> 170793.193 230983.596 100347.700 229401.741 115137.317 440735.543 191086.498 
#>       g232       g233       g234       g235       g236       g237       g238 
#> 301777.286  23139.765 102366.737 388979.107 287613.673 251180.243 118818.249 
#>       g239       g240       g241       g242       g243       g244       g245 
#> 453327.805  82517.048 264955.865  97805.042  74886.988 200987.199 287044.544 
#>       g246       g247       g248       g249       g250       g251       g252 
#> 436717.151 293428.480 368940.544 213897.555 572740.950 414701.821  59818.275 
#>       g253       g254       g255       g256       g257       g258       g259 
#> 222854.910 281719.481 174000.322 404886.209 299839.641  59125.864  38681.211 
#>       g260       g261       g262       g263       g264       g265       g266 
#> 479984.446 399888.586 345371.848 334362.236 141131.281  91741.888 535648.548 
#>       g267       g268       g269       g270       g271       g272       g273 
#> 194884.317 179757.242  62915.028 410102.010  92862.820 494177.779 517817.898 
#>       g274       g275       g276       g277       g278       g279       g280 
#> 312970.306 136691.956 532128.434 138636.472 496833.329 110321.102 243376.899 
#>       g281       g282       g283       g284       g285       g286       g287 
#> 530772.994 379643.350 226119.415  12437.510 266776.566 387497.539 162519.414 
#>       g288       g289       g290       g291       g292       g293       g294 
#>  86986.405 423380.827 197834.125 538221.709   4361.055 450928.687 284678.919 
#>       g295       g296       g297       g298       g299       g300       g301 
#> 257008.239 400096.737 228551.424 135776.384 209274.368  54108.491 539991.440 
#>       g302       g303       g304       g305       g306       g307       g308 
#> 171450.279 156192.942 558729.949 195853.853 360960.074  40053.756 507119.447 
#>       g309       g310       g311       g312       g313       g314       g315 
#> 194691.834 208272.931  27912.523 253976.706 108177.673 378578.861 155476.434 
#>       g316       g317       g318       g319       g320       g321       g322 
#> 537683.752 196425.498 187243.859 248797.020 538134.209 336447.571 119615.102 
#>       g323       g324       g325       g326       g327       g328       g329 
#> 162465.065 124565.942 228946.694 475030.609 114846.726  67248.842   7521.770 
#>       g330       g331       g332       g333       g334       g335       g336 
#>  21868.177   4114.930 299104.844 272773.838   4547.203 247505.820 468705.920 
#>       g337       g338       g339       g340       g341       g342       g343 
#> 384756.888 259368.350  83438.934 173868.887 591146.124 317155.113 584133.189 
#>       g344       g345       g346       g347       g348       g349       g350 
#> 329336.909 357558.924 230280.418   3839.458 346981.956 514548.770 143275.328 
#>       g351       g352       g353       g354       g355       g356       g357 
#> 311901.598 358369.709 111107.281 200060.313 129589.891 247749.917  20761.028 
#>       g358       g359       g360       g361       g362       g363       g364 
#>  85892.285  90046.125  10114.984 335125.362 285164.217  97072.188 407446.114 
#>       g365       g366       g367       g368       g369       g370       g371 
#> 253674.304 157606.692  65821.122 309375.806 401114.419 434552.501 221907.633 
#>       g372       g373       g374       g375       g376       g377       g378 
#> 355635.467 518339.737 401505.577 241858.060 248141.123 130755.601 209862.700 
#>       g379       g380       g381       g382       g383       g384       g385 
#> 137435.060 265364.152 201906.498  52252.354 459955.453 312307.447  66803.537 
#>       g386       g387       g388       g389       g390       g391       g392 
#>  74769.500  81676.768 110418.604 227161.475 153655.730 365629.609 150992.998 
#>       g393       g394       g395       g396       g397       g398       g399 
#> 214425.254 155480.060  34609.899 517612.862 432924.878  28766.338  69573.360 
#>       g400       g401       g402       g403       g404       g405       g406 
#> 277880.482 206648.318 391561.995 170327.157 121863.088 259510.410 336483.562 
#>       g407       g408       g409       g410       g411       g412       g413 
#> 124252.313 301097.899 159724.286  63709.171  64981.236 254882.220  90839.249 
#>       g414       g415       g416       g417       g418       g419       g420 
#> 134508.143 294730.179 211447.429 309375.618 238973.323 472024.032 249821.434 
#>       g421       g422       g423       g424       g425       g426       g427 
#> 349738.741 596856.939  85541.088 517340.530 609012.659 125945.286  80836.792 
#>       g428       g429       g430       g431       g432       g433       g434 
#>  16320.435  66695.781 357377.300 304848.742 293193.810 199200.184 332996.885 
#>       g435       g436       g437       g438       g439       g440       g441 
#> 162030.866 120747.703 176496.894 629589.899 325566.286 113694.705 143504.378 
#>       g442       g443       g444       g445       g446       g447       g448 
#> 177297.510 100118.995 332157.597 376691.209 349221.566 252854.209  10175.329 
#>       g449       g450       g451       g452       g453       g454       g455 
#> 206799.042 341324.994 616636.689 214734.984 177786.738 135205.888 487381.834 
#>       g456       g457       g458       g459       g460       g461       g462 
#> 339435.092 100918.770   9069.378 423482.382 238419.128 439136.604  11293.410 
#>       g463       g464       g465       g466       g467       g468       g469 
#>  36475.707 432490.388 334799.913 151033.475 214387.724 104226.985 273565.917 
#>       g470       g471       g472       g473       g474       g475       g476 
#> 153010.521 447970.805 534788.039 542224.318 270666.161 326204.770 584391.704 
#>       g477       g478       g479       g480       g481       g482       g483 
#> 142657.253 483715.295 562617.604 215461.323 115343.057 402375.063 170698.558 
#>       g484       g485       g486       g487       g488       g489       g490 
#> 112357.418 255970.124 121459.028 122135.824 505176.157 290837.715 318231.676 
#>       g491       g492       g493       g494       g495       g496       g497 
#> 385349.114 185149.285 470136.107 188943.286 361960.272 152396.620 277812.085 
#>       g498       g499       g500       g501       g502       g503       g504 
#>  78949.403 492251.887 349487.544 550378.442 154522.289 177923.737 240387.612 
#>       g505       g506       g507       g508       g509       g510       g511 
#> 120181.935  49357.103 424462.371  61929.308 354871.080 567788.762 100531.610 
#>       g512       g513       g514       g515       g516       g517       g518 
#> 104548.795 378196.996 228000.974 233176.735 132546.179 138019.991 289077.055 
#>       g519       g520       g521       g522       g523       g524       g525 
#> 505375.014  12368.853 137673.991 269538.913 470774.440 163617.176 115367.028 
#>       g526       g527       g528       g529       g530       g531       g532 
#> 111775.222 159126.403 191245.297 326976.308 490135.372 191606.806 535853.403 
#>       g533       g534       g535       g536       g537       g538       g539 
#> 174465.972 150047.724  98887.627 582591.317 115383.143   4450.930  93255.521 
#>       g540       g541       g542       g543       g544       g545       g546 
#>   5529.910   3049.074  66151.350  53201.746 306606.016 550179.991 494699.822 
#>       g547       g548       g549       g550       g551       g552       g553 
#> 302144.736  59776.693 312681.910  31132.956  67888.311 380238.281  45103.910 
#>       g554       g555       g556       g557       g558       g559       g560 
#> 259626.931 282695.387  77425.463 152321.801 193499.397 200319.188 319998.604 
#>       g561       g562       g563       g564       g565       g566       g567 
#> 272099.318 434027.801 173766.594 138084.265  26356.000 254296.448 520038.563 
#>       g568       g569       g570       g571       g572       g573       g574 
#> 169506.601 242526.833 382246.381 536027.887 341465.784 525315.386 551210.125 
#>       g575       g576       g577       g578       g579       g580       g581 
#> 269940.221 169936.242 162486.440 255505.124  62340.765 197448.214 337281.472 
#>       g582       g583       g584       g585       g586       g587       g588 
#> 108463.448 220716.451 261049.896 126151.346 287454.703 101601.268 274005.934 
#>       g589       g590       g591       g592       g593       g594       g595 
#> 190077.424 170602.959 133769.145 237985.125 382224.017 563562.931 524050.698 
#>       g596       g597       g598       g599       g600       g601       g602 
#> 599994.765 569399.813 163829.283 312444.437 144317.995 451572.659 384462.446 
#>       g603       g604       g605       g606       g607       g608       g609 
#> 211032.028 513095.698 372868.871 465268.870 391513.253  95808.982 323389.670 
#>       g610       g611       g612       g613       g614       g615       g616 
#> 199101.885 164038.693 174163.016 192915.856 546824.921 616147.272  45566.452 
#>       g617       g618       g619       g620       g621       g622       g623 
#> 228155.040  36217.899  62633.028 464397.676  78071.313 364514.317 166279.781 
#>       g624       g625       g626       g627       g628       g629       g630 
#>  51060.773 291888.931  53774.636 179071.999 205820.862 278408.730 306829.645 
#>       g631       g632       g633       g634       g635       g636       g637 
#> 426312.510  96695.157 280010.046 465575.963  51450.222 239827.005 164294.749 
#>       g638       g639       g640       g641       g642       g643       g644 
#> 481917.512 218658.378 500017.093  92869.318 184086.092 278720.182 243406.934 
#>       g645       g646       g647       g648       g649       g650       g651 
#> 394438.152 244597.189 279332.232  98561.781 457209.515 356833.738 539544.949 
#>       g652       g653       g654       g655       g656       g657       g658 
#>  71293.604 178177.683 198332.913 227012.222  77612.964 436204.494 270900.179 
#>       g659       g660       g661       g662       g663       g664       g665 
#> 226721.898 164035.478 157317.317 450200.007 201540.821  74435.736 407723.833 
#>       g666       g667       g668       g669       g670       g671       g672 
#> 162964.988 354447.143 244059.157 160887.756 262959.383 330332.715 368916.055 
#>       g673       g674       g675       g676       g677       g678       g679 
#> 263141.124 543717.532 208785.014 116292.290 140935.770 118311.695 504426.314 
#>       g680       g681       g682       g683       g684       g685       g686 
#> 120305.845 127789.063 248003.687 172276.139 159485.047 369450.268 406176.812 
#>       g687       g688       g689       g690       g691       g692       g693 
#> 259103.451 291747.833 357224.810 138312.242 141307.659 369119.465 169245.521 
#>       g694       g695       g696       g697       g698       g699       g700 
#> 165588.378 141454.361 312528.125 254222.412  39239.799 285163.343 148854.240 
#>       g701       g702       g703       g704       g705       g706       g707 
#> 144820.820 450933.050 397527.516 297387.910 119519.742 335199.526 243448.371 
#>       g708       g709       g710       g711       g712       g713       g714 
#> 136357.774 340204.446 161159.295 247347.318 436446.926  92272.232 163617.401 
#>       g715       g716       g717       g718       g719       g720       g721 
#> 225086.812 413754.891 312565.152 544604.710 360184.994 190739.534 399205.683 
#>       g722       g723       g724       g725       g726       g727       g728 
#> 536257.931  17212.360 395304.290 483059.543  64321.878 233941.878 316132.281 
#>       g729       g730       g731       g732       g733       g734       g735 
#> 299521.878 209865.252 386081.934  64069.279 138431.659 285706.988   1127.306 
#>       g736       g737       g738       g739       g740       g741       g742 
#> 610798.176 211416.071 123929.207 419456.168 166088.942 225186.788 475899.328 
#>       g743       g744       g745       g746       g747       g748       g749 
#> 311706.676 166873.801 251088.317 295480.769 106423.091 524269.022  83697.762 
#>       g750       g751       g752       g753       g754       g755       g756 
#> 353366.429 545557.864 532538.773 209965.294 357838.253 162075.666 498598.758 
#>       g757       g758       g759       g760       g761       g762       g763 
#>  77841.232 459056.171 128300.175 500080.092 470727.773 373221.060  83226.211 
#>       g764       g765       g766       g767       g768       g769       g770 
#> 120972.182  82311.189 180500.253  38842.410 110519.852 110970.982 459298.329 
#>       g771       g772       g773       g774       g775       g776       g777 
#> 178063.437 409055.893 111393.679 571331.382 388927.489 428991.878 371791.181 
#>       g778       g779       g780       g781       g782       g783       g784 
#> 157119.272 277311.117 120946.563 617662.221 196491.465 131303.613 188850.769 
#>       g785       g786       g787       g788       g789       g790       g791 
#>  49278.638  90586.457 596505.356 406908.479 163679.968 185244.129 363039.048 
#>       g792       g793       g794       g795       g796       g797       g798 
#> 150397.717 476596.023 192310.999 225181.677 509252.189 405128.532 247548.749 
#>       g799       g800       g801       g802       g803       g804       g805 
#> 173642.152 566132.746  36648.978 535109.996 537060.706 405427.644 136678.177 
#>       g806       g807       g808       g809       g810       g811       g812 
#> 256557.060 510264.515 203219.644 476203.949 203434.885 380041.985 186406.107 
#>       g813       g814       g815       g816       g817       g818       g819 
#> 138903.354 171760.349 234028.253 454552.091 345477.491 137250.306 331991.263 
#>       g820       g821       g822       g823       g824       g825       g826 
#> 319917.856 502079.726 171384.394 293918.820 253238.530 356743.337 536095.598 
#>       g827       g828       g829       g830       g831       g832       g833 
#>  80873.797  80273.120 194099.737 317669.408 314257.408 554900.211 118019.047 
#>       g834       g835       g836       g837       g838       g839       g840 
#>  98273.898 468360.673 168538.479 174783.899 382856.824 554941.770 200330.133 
#>       g841       g842       g843       g844       g845       g846       g847 
#> 318581.510 266154.512 515652.523 464301.461 101620.156 126226.031 299051.677 
#>       g848       g849       g850       g851       g852       g853       g854 
#>  93591.813 169825.279 218797.185 544478.332 298952.371 123709.182 485546.530 
#>       g855       g856       g857       g858       g859       g860       g861 
#> 183126.704 187853.268 332768.594  57371.351 408292.532 471380.484 144829.562 
#>       g862       g863       g864       g865       g866       g867       g868 
#> 236862.778 201163.736 461839.214 506836.027  71216.451 214545.892 114628.773 
#>       g869       g870       g871       g872       g873       g874       g875 
#> 182886.576 269326.423 124194.520 265310.217 403476.102 563232.336 541931.297 
#>       g876       g877       g878       g879       g880       g881       g882 
#> 105059.100 462968.318 105256.586 117713.478 153842.845 527585.240 161426.168 
#>       g883       g884       g885       g886       g887       g888       g889 
#> 475183.878 546871.945 162332.122  43654.820 137654.687  86981.576 299882.184 
#>       g890       g891       g892       g893       g894       g895       g896 
#> 126300.870 199337.788 459939.908 122345.197 477963.752 231244.597 241847.870 
#>       g897       g898       g899       g900       g901       g902       g903 
#> 159358.865 267216.322 259946.770 353858.666 115147.168 382759.512 383791.942 
#>       g904       g905       g906       g907       g908       g909       g910 
#> 141901.242 361256.562 411573.835 107767.120 378230.953 110344.770 592971.217 
#>       g911       g912       g913       g914       g915       g916       g917 
#> 343872.014 540313.351  93068.300  83300.749 471552.472 521094.687  53637.171 
#>       g918       g919       g920       g921       g922       g923       g924 
#> 167751.408  81157.189 135415.852  87618.779 295804.881 225452.668 518929.448 
#>       g925       g926       g927       g928       g929       g930       g931 
#> 545243.956 324230.629 181590.939  38340.386 447893.835 488297.668 124560.672 
#>       g932       g933       g934       g935       g936       g937       g938 
#> 119437.555 308797.123 148507.136 376686.779 483045.157 425594.433  72320.032 
#>       g939       g940       g941       g942       g943       g944       g945 
#> 157852.668 120882.747 454281.393 144176.629  17108.968 138750.403 562540.426 
#>       g946       g947       g948       g949       g950       g951       g952 
#>  42972.217  84615.539 458565.668 418972.684 277888.498 408905.020 299180.885 
#>       g953       g954       g955       g956       g957       g958       g959 
#> 175805.498 287085.354 462068.968 451116.182 558908.842  83273.910 438152.240 
#>       g960       g961       g962       g963       g964       g965       g966 
#>  86680.982 572175.974 243740.862 460083.871 394914.504  91773.618 543202.726 
#>       g967       g968       g969       g970       g971       g972       g973 
#> 216906.143 130390.683 468644.727 349813.928 221933.254 172916.724 230420.489 
#>       g974       g975       g976       g977       g978       g979       g980 
#> 588619.942 604567.405 557583.934 447921.595 541061.446 495315.494 287195.243 
#>       g981       g982       g983       g984       g985       g986       g987 
#> 134368.478 417576.663 500393.385 152698.558  71438.262 156708.917 260171.245 
#>       g988       g989       g990       g991       g992       g993       g994 
#> 383520.188 352781.328  71958.369 237402.607  12099.091 386025.784 481422.677 
#>       g995       g996       g997       g998       g999      g1000      g1001 
#> 312370.426  52684.797 396121.114 547936.128 139487.909 135898.633 241027.360 
#>      g1002      g1003      g1004      g1005      g1006      g1007      g1008 
#> 217382.765 168747.143 107723.295 167894.231 137522.861 390690.953 114894.415 
#>      g1009      g1010      g1011      g1012      g1013      g1014      g1015 
#> 255016.746 476999.949 164463.779 197405.338 176085.766 188557.046 407046.770 
#>      g1016      g1017      g1018      g1019      g1020      g1021      g1022 
#> 125269.027 507160.134 370205.898 454519.673 443808.437 553531.634 257767.781 
#>      g1023      g1024      g1025      g1026      g1027      g1028      g1029 
#> 465754.814 582331.197 305135.629 368622.046 325578.165 457263.177 569350.381 
#>      g1030      g1031      g1032      g1033      g1034      g1035      g1036 
#>  91367.300 370454.125 151994.512 391095.765 252005.039 293300.623 146243.756 
#>      g1037      g1038      g1039      g1040      g1041      g1042      g1043 
#> 164816.350 130589.895 465542.805 363203.354 538956.710 379734.576  92842.504 
#>      g1044      g1045      g1046      g1047      g1048      g1049      g1050 
#>  75083.576 378008.007 320367.401  74193.878 249421.899 221025.134 565476.573 
#>      g1051      g1052      g1053      g1054      g1055      g1056      g1057 
#>  69802.189 347196.221 319285.409  50491.698 323356.060  46269.908 421744.184 
#>      g1058      g1059      g1060      g1061      g1062      g1063      g1064 
#> 168725.361 236130.586 102378.869 554658.236 514239.781 124878.999 393667.189 
#>      g1065      g1066      g1067      g1068      g1069      g1070      g1071 
#>   1602.236 399784.351 488519.945 219621.133 260016.232 617235.770 256624.691 
#>      g1072      g1073      g1074      g1075      g1076      g1077      g1078 
#>  77510.156 245080.483 583461.737 209444.337 377581.968 486631.853 182998.349 
#>      g1079      g1080      g1081      g1082      g1083      g1084      g1085 
#> 657348.778 385692.504 150782.973 349693.249 273446.691 584720.756 222980.174 
#>      g1086      g1087      g1088      g1089      g1090      g1091      g1092 
#>  76502.184 398811.735 545436.082 121276.899 499066.802 191794.673 108620.598 
#>      g1093      g1094      g1095      g1096      g1097      g1098      g1099 
#> 405925.737 363809.472 107703.795 291183.313 243868.133 204368.691 351826.242 
#>      g1100      g1101      g1102      g1103      g1104      g1105      g1106 
#> 191013.569 276324.673 330645.687 165332.090 115044.398 342077.512 359272.825 
#>      g1107      g1108      g1109      g1110      g1111      g1112      g1113 
#> 175618.358 144775.336  46545.275 249566.606 140559.157 166692.898 255515.457 
#>      g1114      g1115      g1116      g1117      g1118      g1119      g1120 
#> 382540.014  87623.047 518140.977  95261.275 431587.691 345553.668 260968.154 
#>      g1121      g1122      g1123      g1124      g1125      g1126      g1127 
#> 422066.401 576329.346 452352.964 165092.097 398502.504 272058.107 129901.833 
#>      g1128      g1129      g1130      g1131      g1132      g1133      g1134 
#> 448534.749 470428.295 406216.186 361565.581 216606.286 219098.418 135290.240 
#>      g1135      g1136      g1137      g1138      g1139      g1140      g1141 
#> 305878.586 188536.623 139685.577 116755.242 459926.349  94241.393 632497.649 
#>      g1142      g1143      g1144      g1145      g1146      g1147      g1148 
#> 339406.536 411382.236 453930.903 132769.222 181972.581 145898.442 303304.088 
#>      g1149      g1150      g1151      g1152      g1153      g1154      g1155 
#> 146177.091  72088.297 429078.853 373799.166 115317.023 563046.629 140470.644 
#>      g1156      g1157      g1158      g1159      g1160      g1161      g1162 
#> 153218.663  89224.530 425468.440 331414.866 125488.951 198048.832 521898.421 
#>      g1163      g1164      g1165      g1166      g1167      g1168      g1169 
#> 191459.895 109178.256 147668.733 421274.053 377493.092 426215.928 424350.829 
#>      g1170      g1171      g1172      g1173      g1174      g1175      g1176 
#> 295594.607 193091.601 373929.733  32535.120 261090.862 332357.188  75693.372 
#>      g1177      g1178      g1179      g1180      g1181      g1182      g1183 
#> 604200.853 168540.297 229304.676 226149.668 474955.824 419491.544 440984.887 
#>      g1184      g1185      g1186      g1187      g1188      g1189      g1190 
#> 434571.179  14008.209 156630.340 276443.977 138271.806 275014.010 273786.128 
#>      g1191      g1192      g1193      g1194      g1195      g1196      g1197 
#>  95588.133 187454.604 523142.165 256323.113 182900.093 175286.410 452167.638 
#>      g1198      g1199      g1200      g1201      g1202      g1203      g1204 
#> 212508.388 542007.996 129972.770 124747.962 183682.867 626347.272 137550.433 
#>      g1205      g1206      g1207      g1208      g1209      g1210      g1211 
#> 314368.429  93241.500 400894.092 297363.539 180083.152 101552.813 306904.221 
#>      g1212      g1213      g1214      g1215      g1216      g1217      g1218 
#> 132807.193 485839.186 146748.042 159058.415 116175.527 592116.055 350890.544 
#>      g1219      g1220      g1221      g1222      g1223      g1224      g1225 
#> 542503.199 121454.918 150773.184 521044.279 547235.275 543688.369 188406.200 
#>      g1226      g1227      g1228      g1229      g1230      g1231      g1232 
#> 115190.028 288610.107 131941.548 207570.133 102072.394  74069.700 137317.179 
#>      g1233      g1234      g1235      g1236      g1237      g1238      g1239 
#> 440232.415  74157.500 311049.360 116144.221 318541.315  64166.627 565319.544 
#>      g1240      g1241      g1242      g1243      g1244      g1245      g1246 
#> 112035.197 521182.918 442551.985  72253.217 145145.704 231270.440 246177.014 
#>      g1247      g1248      g1249      g1250      g1251      g1252      g1253 
#> 437687.894 470527.128  54736.336  83004.004  73660.111 433826.281  95154.337 
#>      g1254      g1255      g1256      g1257      g1258      g1259      g1260 
#> 208700.608 179527.747 121417.028 290900.921 510157.362 122005.053 409411.799 
#>      g1261      g1262      g1263      g1264      g1265      g1266      g1267 
#> 157776.594 188712.281 217845.544 162282.570 158461.630 130587.182 150212.801 
#>      g1268      g1269      g1270      g1271      g1272      g1273      g1274 
#> 449323.426  93437.738 463042.507 419539.189 543214.079 293641.088  58834.562 
#>      g1275      g1276      g1277      g1278      g1279      g1280      g1281 
#> 379402.770 213917.571 325963.184  23761.304 435272.805 621843.827 372258.833 
#>      g1282      g1283      g1284      g1285      g1286      g1287      g1288 
#> 504005.032 229556.570 475462.749 109083.300 322646.351 175006.100 250391.189 
#>      g1289      g1290      g1291      g1292      g1293      g1294      g1295 
#> 464641.921 126194.451 234247.894 527694.738 151641.633  93751.225 475067.088 
#>      g1296      g1297      g1298      g1299      g1300      g1301      g1302 
#> 128907.179 108837.608 295624.641 277212.338 490940.221 249769.947 244422.217 
#>      g1303      g1304      g1305      g1306      g1307      g1308      g1309 
#> 114712.644 518345.498 121668.437 358271.896 430986.594  76758.350 403648.566 
#>      g1310      g1311      g1312      g1313      g1314      g1315      g1316 
#> 529036.726 316072.031 367776.733 327474.848 173020.661  60019.555 103829.767 
#>      g1317      g1318      g1319      g1320      g1321      g1322      g1323 
#> 480817.902 585932.899 244587.853 134365.429 432436.834 355476.949 510304.360 
#>      g1324      g1325      g1326      g1327      g1328      g1329      g1330 
#> 281617.834 184454.092 206430.329 473736.447 117100.913  88855.071 180391.786 
#>      g1331      g1332      g1333      g1334      g1335      g1336      g1337 
#> 136172.641  57208.017 326212.177 265315.475 203674.973 138492.627 253913.480 
#>      g1338      g1339      g1340      g1341      g1342      g1343      g1344 
#> 113089.315 574690.795 537542.221 335074.829 515397.125  84033.539  48979.853 
#>      g1345      g1346      g1347      g1348      g1349      g1350      g1351 
#> 289657.884 144076.046  99683.128 349945.461 124216.060  90762.433 339714.111 
#>      g1352      g1353      g1354      g1355      g1356      g1357      g1358 
#>  81993.870 389746.693 249659.182 509933.587 234344.749 561218.813 120170.608 
#>      g1359      g1360      g1361      g1362      g1363      g1364      g1365 
#> 301535.576 171038.369 489168.429 517914.834  38769.440 142757.931 434250.165 
#>      g1366      g1367      g1368      g1369      g1370      g1371      g1372 
#> 199143.902  99653.397 152705.318 377540.790 226107.112 392668.749 374080.727 
#>      g1373      g1374      g1375      g1376      g1377      g1378      g1379 
#> 163558.620 485832.311 174361.252 102165.962 422592.442  97288.275 512360.092 
#>      g1380      g1381      g1382      g1383      g1384      g1385      g1386 
#> 180530.157 124521.111 517161.122 434402.133 185169.418 321705.888 223024.028 
#>      g1387      g1388      g1389      g1390      g1391      g1392      g1393 
#> 122590.351 258874.318 444546.991 230247.704 294951.788 134814.912 195165.306 
#>      g1394      g1395      g1396      g1397      g1398      g1399      g1400 
#> 464433.458 528345.032 115546.386 372142.340 153743.143 339108.286  81748.096 
#>      g1401      g1402      g1403      g1404      g1405      g1406      g1407 
#> 239926.487 454203.236 366183.232 255758.756 164109.698  86870.387 186261.318 
#>      g1408      g1409      g1410      g1411      g1412      g1413      g1414 
#> 240328.442 253796.746  94242.194  81194.876 118427.265  90067.318 344287.829 
#>      g1415      g1416      g1417      g1418      g1419      g1420      g1421 
#> 103769.410 172388.537 411625.698 498036.802 379884.832 149552.369 518323.203 
#>      g1422      g1423      g1424      g1425      g1426      g1427      g1428 
#> 268836.641 165627.562 473987.824 213449.674 443515.719 386575.090 161862.705 
#>      g1429      g1430      g1431      g1432      g1433      g1434      g1435 
#> 142900.285 501829.071 569325.956 198207.666 178723.189 321236.902 477607.023 
#>      g1436      g1437      g1438      g1439      g1440      g1441      g1442 
#> 415709.749 205739.940  63964.318 370002.039 218692.025 513269.963 143585.820 
#>      g1443      g1444      g1445      g1446      g1447      g1448      g1449 
#> 159969.304 495276.005 163695.941 263281.737 359136.532  76340.677 451201.452 
#>      g1450      g1451      g1452      g1453      g1454      g1455      g1456 
#> 217821.383 617169.984 527722.607 222621.016 407197.872 367589.487 154426.684 
#>      g1457      g1458      g1459      g1460      g1461      g1462      g1463 
#> 101641.647 457220.564 335703.172 369727.594 267106.574 324009.991 550902.930 
#>      g1464      g1465      g1466      g1467      g1468      g1469      g1470 
#> 154236.088  96836.111 432239.410 505479.311 465925.521 453873.523 173748.641 
#>      g1471      g1472      g1473      g1474      g1475      g1476      g1477 
#> 230872.855 170729.461 354816.175 269967.819  96836.391 500426.652 185494.448 
#>      g1478      g1479      g1480      g1481      g1482      g1483      g1484 
#> 454452.031 198571.196 207721.806 290452.254  73283.221  87918.534  42537.171 
#>      g1485      g1486      g1487      g1488      g1489      g1490      g1491 
#> 541017.844 219003.382 132906.827 503053.566 443031.315 207333.726 328187.349 
#>      g1492      g1493      g1494      g1495      g1496      g1497      g1498 
#> 331963.547  37839.666 375045.820 104713.093 428223.472 367075.601 389342.203 
#>      g1499      g1500      g1501      g1502      g1503      g1504      g1505 
#>  90584.881 365658.469  41473.784 121835.788  30152.332 108351.461  86808.695 
#>      g1506      g1507      g1508      g1509      g1510      g1511      g1512 
#> 136698.429 383987.797 291661.919 554662.336 558799.275 505570.620 367749.676 
#>      g1513      g1514      g1515      g1516      g1517      g1518      g1519 
#> 140662.829 137700.613  96516.313 255904.564 448828.694 295566.113 593893.225 
#>      g1520      g1521      g1522      g1523      g1524      g1525      g1526 
#> 446312.564 103701.530 110097.590  94406.014 395815.355 515310.797 444284.120 
#>      g1527      g1528      g1529      g1530      g1531      g1532      g1533 
#> 485148.114 539903.767 332342.023 429544.214 165974.263 479234.681 362051.211 
#>      g1534      g1535      g1536      g1537      g1538      g1539      g1540 
#> 155343.124 219396.336 141736.221 463657.383 436444.562 324060.350  60407.257 
#>      g1541      g1542      g1543      g1544      g1545      g1546      g1547 
#> 275766.480 505318.507 216189.956 314100.802 126030.899 144250.752  30029.371 
#>      g1548      g1549      g1550      g1551      g1552      g1553      g1554 
#> 472591.756 359077.824 309315.039 187802.207 112045.827 152293.833  74721.824 
#>      g1555      g1556      g1557      g1558      g1559      g1560      g1561 
#> 541767.475  66560.779 165103.705 225464.156  82484.281  33909.440 232222.758 
#>      g1562      g1563      g1564      g1565      g1566      g1567      g1568 
#> 373680.959  74121.762 110737.075  92734.504 193778.838 206148.636 393743.876 
#>      g1569      g1570      g1571      g1572      g1573      g1574      g1575 
#> 347829.733 117752.759 179298.023   1529.530  70921.551  70790.254 337093.544 
#>      g1576      g1577      g1578      g1579      g1580      g1581      g1582 
#> 481900.290 230678.898 495786.071 167714.017 418874.687 206241.957 621135.451 
#>      g1583      g1584      g1585      g1586      g1587      g1588      g1589 
#> 102061.221 109775.091 279690.252 175529.668 387786.931 516439.949 242959.715 
#>      g1590      g1591      g1592      g1593      g1594      g1595      g1596 
#> 316588.577 315248.354 200764.143 102597.636 237658.226 308976.609 583036.142 
#>      g1597      g1598      g1599      g1600      g1601      g1602      g1603 
#> 349717.152 293707.738 138230.023  24443.504 324728.107 458108.999 413263.060 
#>      g1604      g1605      g1606      g1607      g1608      g1609      g1610 
#> 181962.272  76360.103 259407.710 185444.788 484893.581 370818.414 182953.541 
#>      g1611      g1612      g1613      g1614      g1615      g1616      g1617 
#> 574500.203 395225.853 270566.555   6492.145   7045.014 472824.649  90103.716 
#>      g1618      g1619      g1620      g1621      g1622      g1623      g1624 
#> 460240.447  43748.160 452985.609 542294.423 583054.211 500605.175 445521.060 
#>      g1625      g1626      g1627      g1628      g1629      g1630      g1631 
#> 420459.516 403288.570 247443.014 467419.888 312137.918 295581.275  25290.950 
#>      g1632      g1633      g1634      g1635      g1636      g1637      g1638 
#>   9345.304 527817.640 391485.848  96124.476 281416.220  82564.598 288732.107 
#>      g1639      g1640      g1641      g1642      g1643      g1644      g1645 
#> 194668.598 107346.903 128401.715 238038.479 211038.641 534836.746 363250.378 
#>      g1646      g1647      g1648      g1649      g1650      g1651      g1652 
#> 124607.931 134603.848 517286.659 435904.276 550312.543 520958.926 229121.171 
#>      g1653      g1654      g1655      g1656      g1657      g1658      g1659 
#> 119970.853 423937.211 214560.806 240693.253 341473.214 354744.189 511056.134 
#>      g1660      g1661      g1662      g1663      g1664      g1665      g1666 
#> 327519.974 483580.047 247780.217 130803.286 216650.544 444355.161  15413.887 
#>      g1667      g1668      g1669      g1670      g1671      g1672      g1673 
#> 303178.351 215717.096 184508.225 207946.143 360926.988 381715.627 138589.060 
#>      g1674      g1675      g1676      g1677      g1678      g1679      g1680 
#> 278564.435  99577.693 490353.317 113539.236 513508.998 142346.781 463502.907 
#>      g1681      g1682      g1683      g1684      g1685      g1686      g1687 
#> 158159.157 495774.057 358384.232 117824.408 432049.630 366540.429 467475.643 
#>      g1688      g1689      g1690      g1691      g1692      g1693      g1694 
#> 183155.563 343495.059 439946.977 163073.279 247300.586 130201.768  86839.887 
#>      g1695      g1696      g1697      g1698      g1699      g1700      g1701 
#> 127706.706  73692.700 123632.633 281128.903 222437.252 349082.318 250366.437 
#>      g1702      g1703      g1704      g1705      g1706      g1707      g1708 
#>  88854.382 103038.452 355572.360 266877.313 267839.645 305410.834 285781.194 
#>      g1709      g1710      g1711      g1712      g1713      g1714      g1715 
#> 187680.788 133278.995 310708.281 247026.859 152002.590 498210.974 474365.179 
#>      g1716      g1717      g1718      g1719      g1720      g1721      g1722 
#> 456409.959 161771.509 395384.833 328403.207 185263.704 222714.000 135857.629 
#>      g1723      g1724      g1725      g1726      g1727      g1728      g1729 
#> 207405.222 144453.268 115315.738 302472.046 281898.484 150935.512 515250.921 
#>      g1730      g1731      g1732      g1733      g1734      g1735      g1736 
#> 558154.194 135822.784 293461.318 143444.337 156604.117  91435.383 102942.584 
#>      g1737      g1738      g1739      g1740      g1741      g1742      g1743 
#>  26989.902 430511.189 275078.741 100424.604  99237.756  71503.063 104598.372 
#>      g1744      g1745      g1746      g1747      g1748      g1749      g1750 
#> 139868.423 101681.512 153122.010 129527.926 537189.552 178783.346 334035.437 
#>      g1751      g1752      g1753      g1754      g1755      g1756      g1757 
#> 350924.415 199537.696 311544.896 313481.555 287216.257 101190.096  62459.143 
#>      g1758      g1759      g1760      g1761      g1762      g1763      g1764 
#> 213137.951 267858.059 220850.028  73282.447 115854.691 110266.795 150811.096 
#>      g1765      g1766      g1767      g1768      g1769      g1770      g1771 
#>  17041.659 263122.354 351523.736 108426.829 386097.423 416933.480 186110.935 
#>      g1772      g1773      g1774      g1775      g1776      g1777      g1778 
#> 425361.337 332837.082   2625.158 103296.447 175463.692 143188.433 127797.995 
#>      g1779      g1780      g1781      g1782      g1783      g1784      g1785 
#> 158924.735 540452.921   1793.982 124070.620 377462.683 488634.286 181438.475 
#>      g1786      g1787      g1788      g1789      g1790      g1791      g1792 
#> 135400.752 134078.328 146024.412 162859.498 470129.959 432281.106 185939.433 
#>      g1793      g1794      g1795      g1796      g1797      g1798      g1799 
#> 498658.489 532170.381 431757.243 397955.838 232437.160 140603.586 281034.753 
#>      g1800      g1801      g1802      g1803      g1804      g1805      g1806 
#> 166472.088 238151.956 247336.386 222557.663 160186.562 421440.218 219735.092 
#>      g1807      g1808      g1809      g1810      g1811      g1812      g1813 
#> 181973.401 208670.318 359530.691 260654.028 449541.581 103818.752 550009.687 
#>      g1814      g1815      g1816      g1817      g1818      g1819      g1820 
#> 305386.587 105268.125  65758.613 218939.254 406834.899 344848.085 472217.189 
#>      g1821      g1822      g1823      g1824      g1825      g1826      g1827 
#> 286264.724 393592.693 141490.801  56415.700  94942.586   7178.209 273172.520 
#>      g1828      g1829      g1830      g1831      g1832      g1833      g1834 
#> 206176.867 126318.253  85341.898 405309.924 296243.350  88044.676 427050.523 
#>      g1835      g1836      g1837      g1838      g1839      g1840      g1841 
#> 220332.982 455998.125 487785.388 122165.834 374848.543 287507.382 223907.623 
#>      g1842      g1843      g1844      g1845      g1846      g1847      g1848 
#>  54998.985 562578.700 442895.351 413402.934  72779.985 418072.231 186200.457 
#>      g1849      g1850      g1851      g1852      g1853      g1854      g1855 
#> 234449.074 143768.715 559240.530 516638.848 197878.724 164571.623 119231.156 
#>      g1856      g1857      g1858      g1859      g1860      g1861      g1862 
#> 142345.673 332324.921 513797.738 441880.934 256882.242 101523.415 390192.447 
#>      g1863      g1864      g1865      g1866      g1867      g1868      g1869 
#> 134638.120 250928.918 541773.350 160041.792 551689.150 224775.784 126990.225 
#>      g1870      g1871      g1872      g1873      g1874      g1875      g1876 
#> 412838.926 126820.294 483850.290 522185.129 509626.704  94909.684 447006.885 
#>      g1877      g1878      g1879      g1880      g1881      g1882      g1883 
#> 227004.288  92535.866 469610.885 173085.482 198059.489 119932.240 479592.143 
#>      g1884      g1885      g1886      g1887      g1888      g1889      g1890 
#> 282720.134 524270.290 383020.607 249231.494 376008.951 520986.521 293353.189 
#>      g1891      g1892      g1893      g1894      g1895      g1896      g1897 
#> 509149.175 431086.594 155489.214 291744.629 305329.442 203022.926   1169.543 
#>      g1898      g1899      g1900      g1901      g1902      g1903      g1904 
#> 274705.318 526007.381 311537.297 515829.853  21727.515 173573.451 493510.942 
#>      g1905      g1906      g1907      g1908      g1909      g1910      g1911 
#> 174719.630 175498.822 165696.401 412107.016 234022.401 184316.507 202221.544 
#>      g1912      g1913      g1914      g1915      g1916      g1917      g1918 
#> 359521.186 480317.691 405698.027 292715.350 236548.770 223802.415  68308.876 
#>      g1919      g1920      g1921      g1922      g1923      g1924      g1925 
#> 130612.853 481247.175 401944.967 271115.994 123249.552 285799.571 121485.028 
#>      g1926      g1927      g1928      g1929      g1930      g1931      g1932 
#>  92008.673 360741.311 142300.706 399255.429 387726.117 235363.795 541189.885 
#>      g1933      g1934      g1935      g1936      g1937      g1938      g1939 
#>  87683.894 125322.480 173967.742  51726.410 257170.544 142093.435 256939.263 
#>      g1940      g1941      g1942      g1943      g1944      g1945      g1946 
#> 170384.661 418084.146 454966.257  36043.401 548710.584 594333.361 501718.705 
#>      g1947      g1948      g1949      g1950      g1951      g1952      g1953 
#> 364685.268 176002.705 465775.159 448332.802 397711.475 481629.691 404643.715 
#>      g1954      g1955      g1956      g1957      g1958      g1959      g1960 
#> 555971.380 461577.020 286319.467 294422.859 118991.820 264063.894 387469.088 
#>      g1961      g1962      g1963      g1964      g1965      g1966      g1967 
#> 240060.010 494896.418 545518.465  78657.571 219137.865 387505.541 494677.673 
#>      g1968      g1969      g1970      g1971      g1972      g1973      g1974 
#> 237229.641 103878.171 136700.784 235323.507 504021.968 203628.207 156771.209 
#>      g1975      g1976      g1977      g1978      g1979      g1980      g1981 
#> 293113.410  74816.448 325384.401 350360.403 124857.128 380443.221 437249.032 
#>      g1982      g1983      g1984      g1985      g1986      g1987      g1988 
#> 213511.491  84993.971 125469.742 394145.510 569337.163 147797.762  90656.152 
#>      g1989      g1990      g1991      g1992      g1993      g1994      g1995 
#> 174671.348 308677.982 228327.903 272976.548 218848.523 105391.577  71132.253 
#>      g1996      g1997      g1998      g1999      g2000      g2001      g2002 
#> 181400.451 210039.901 149283.850 163951.315 126201.878 247479.134 150022.512 
#>      g2003      g2004      g2005      g2006      g2007      g2008      g2009 
#>  51508.122 184797.182 179115.251 170910.507 457936.343 126005.917 142402.604 
#>      g2010      g2011      g2012      g2013      g2014      g2015      g2016 
#> 144341.623 201544.530 203315.429  71215.618 172035.577 204702.467 159707.774 
#>      g2017      g2018      g2019      g2020      g2021      g2022      g2023 
#> 191750.249 172484.695 149548.447 436896.977 126603.848 469617.645  74817.189 
#>      g2024      g2025      g2026      g2027      g2028      g2029      g2030 
#> 104264.435 188388.927 134720.899  56397.401 158428.752 176953.134  94937.834 
#>      g2031      g2032      g2033      g2034      g2035      g2036      g2037 
#>  34266.899 215205.232 200542.446 152159.194 159881.835 104489.809 238673.504 
#>      g2038      g2039      g2040      g2041      g2042      g2043      g2044 
#> 120479.959 107109.065 111680.297 102411.574 436069.133  96859.349  44809.942 
#>      g2045      g2046      g2047      g2048      g2049      g2050      g2051 
#>  72047.221 237900.179 132159.103  49663.640 153139.950 355711.706 119447.871 
#>      g2052      g2053      g2054      g2055      g2056      g2057      g2058 
#>  55958.498 511780.555 184287.534 395198.123  50081.114 119967.616 100893.779 
#>      g2059      g2060      g2061      g2062      g2063      g2064      g2065 
#> 126025.958 322137.305 347139.781 102300.805 132267.124 198615.465 296118.338 
#>      g2066      g2067      g2068      g2069      g2070      g2071      g2072 
#> 329103.607 109453.226  92062.563  38518.659 124494.806  37050.318  95317.152 
#>      g2073      g2074      g2075      g2076      g2077      g2078      g2079 
#> 397347.451  76805.687 272738.566  27238.512 251191.893  17139.648  42586.189 
#>      g2080      g2081      g2082      g2083      g2084      g2085      g2086 
#>  95319.317 174404.910  70492.120 261140.931 122401.125 308472.845 152200.980 
#>      g2087      g2088      g2089      g2090      g2091      g2092      g2093 
#> 195448.128 201484.687 189767.281  55136.328 449485.060 173118.649 283331.683 
#>      g2094      g2095      g2096      g2097      g2098      g2099      g2100 
#>  99311.090  41720.959  43558.189 179817.124 207185.249 522715.501 106640.655 
#>      g2101      g2102      g2103      g2104      g2105      g2106      g2107 
#> 101595.350 401571.281 119324.433 222755.211 151903.343 135908.573  56410.716 
#>      g2108      g2109      g2110      g2111      g2112      g2113      g2114 
#>  91199.369 187383.781 191795.888 280507.971 179612.673 503111.738 140078.145 
#>      g2115      g2116      g2117      g2118      g2119      g2120      g2121 
#> 167817.531 299837.461 565196.865 116678.790 151096.498 347340.150 180027.143 
#>      g2122      g2123      g2124      g2125      g2126      g2127      g2128 
#>  87356.060 411641.576 191525.833 192649.397 110707.504 338954.642  53625.085 
#>      g2129      g2130      g2131      g2132      g2133      g2134      g2135 
#>  76156.627 244018.461  82943.475 147064.666  76298.924 159261.337 380867.961 
#>      g2136      g2137      g2138      g2139      g2140      g2141      g2142 
#> 153027.705 212920.792 269170.249 563071.241 310035.467  48482.232 152540.681 
#>      g2143      g2144      g2145      g2146      g2147      g2148      g2149 
#> 105530.317 209807.636  94287.426 218031.123 106810.387 316595.876 197164.820 
#>      g2150      g2151      g2152      g2153      g2154      g2155      g2156 
#> 418281.175  95183.891 206783.935 384356.313  53794.894 229181.275 219376.829 
#>      g2157      g2158      g2159      g2160      g2161      g2162      g2163 
#> 441428.539 157364.193 179326.562 342007.078 251864.848 400100.710  64142.982 
#>      g2164      g2165      g2166      g2167      g2168      g2169      g2170 
#> 277516.931 435650.480 371023.046  88266.580 180496.663 134237.206 241007.397 
#>      g2171      g2172      g2173      g2174      g2175      g2176      g2177 
#> 432835.718 298098.899 404587.931 323931.442 230882.489 211412.401  59518.070 
#>      g2178      g2179      g2180      g2181      g2182      g2183      g2184 
#> 329156.738 297207.318 400199.604  60487.724 283334.526 106792.713 188401.323 
#>      g2185      g2186      g2187      g2188      g2189      g2190      g2191 
#> 120452.498 219255.284  81048.405 372175.301  87301.633  63323.405 397786.730 
#>      g2192      g2193      g2194      g2195      g2196      g2197      g2198 
#> 278551.774 168041.676 295481.004 352041.650 152469.677 356541.189 316625.999 
#>      g2199      g2200      g2201      g2202      g2203      g2204      g2205 
#> 162909.304 225854.532 219922.272  73018.602 133640.618 174177.068 367261.046 
#>      g2206      g2207      g2208      g2209      g2210      g2211      g2212 
#> 505650.878 169832.931  99995.295 214838.917 361192.988 300395.802 242475.378 
#>      g2213      g2214      g2215      g2216      g2217      g2218      g2219 
#> 451677.878 409137.160 562426.217 257221.017 223339.913 556160.539 246203.031 
#>      g2220      g2221      g2222      g2223      g2224      g2225      g2226 
#> 474879.028 578985.336 228538.074 384006.222  65938.945 324782.145 157410.403 
#>      g2227      g2228      g2229      g2230      g2231      g2232      g2233 
#> 161479.566 447666.286 256594.590 378529.752 410287.184 190831.394 159405.977 
#>      g2234      g2235      g2236      g2237      g2238      g2239      g2240 
#> 592293.152 224455.770  86349.382 238778.243 190576.963 317712.010 308369.627 
#>      g2241      g2242      g2243      g2244      g2245      g2246      g2247 
#> 428954.907 133418.709 113188.555 203096.928   4292.498  79111.491 137883.048 
#>      g2248      g2249      g2250      g2251      g2252      g2253      g2254 
#> 221369.378 513906.308 145445.092 143526.156 225988.221 236119.694  93743.159 
#>      g2255      g2256      g2257      g2258      g2259      g2260      g2261 
#> 412599.756 543710.365  98104.908 277875.544 202256.834 166727.380 261278.683 
#>      g2262      g2263      g2264      g2265      g2266      g2267      g2268 
#> 108063.956 100946.483 163143.007 202527.231 172942.865 102323.580 215555.378 
#>      g2269      g2270      g2271      g2272      g2273      g2274      g2275 
#>  49386.917 466802.673 139458.684 440081.318 429349.670 419853.242 138532.392 
#>      g2276      g2277      g2278      g2279      g2280      g2281      g2282 
#> 355630.111 192473.272 107886.534 293438.236 315971.719 368243.865 137610.789 
#>      g2283      g2284      g2285      g2286      g2287      g2288      g2289 
#> 450758.219 146739.272 295453.461  91627.201 423711.676 197151.609 215396.265 
#>      g2290      g2291      g2292      g2293      g2294      g2295      g2296 
#> 313222.791 245429.019 611681.538 169581.985 168842.935 270175.313 131815.215 
#>      g2297      g2298      g2299      g2300      g2301      g2302      g2303 
#> 127145.150 213170.208 378303.272 372223.555 151174.240 219465.405 358926.645 
#>      g2304      g2305      g2306      g2307      g2308 
#>  97453.046 133525.175 142097.608 289027.313  74678.113